Consolidate duplicated testsand change ingest endpoint success response code to 202
